Fix build on 10.8 (tested module load, not tested on older OSX or Linux).
The Asterisk headers use the AST_INLINE_API macro to declare helper
functions. For some reason the default is to multiply implement
everything (non-static!). For some reason this worked previously but
doesn't now so use the LOW_MEMORY macro to force it to only declare
the functions but not the implementations.